AI_Detect/requirements.txt
fastapi>=0.110
uvicorn>=0.29
scikit-learn>=1.3
numpy>=1.24
joblib>=1.3
pydantic>=2.0
requests>=2.31
PyYAML>=6.0
